src/AppLogistics.Controllers/Configuration/Sexes/SexesController.cs
using AppLogistics.Components.Mvc;
using AppLogistics.Objects;
using AppLogistics.Services;
using AppLogistics.Validators;
using Microsoft.AspNetCore.Mvc;

namespace AppLogistics.Controllers.Configuration
{
    [Area("Configuration")]
    public class SexesController : ValidatedController<ISexValidator, ISexService>
    {
        public SexesController(ISexValidator validator, ISexService service)
            : base(validator, service)
        {
        }

        [HttpGet]
        public ViewResult Index()
        {
            return View(Service.GetViews());
        }

        [HttpGet]
        public ViewResult Create()
        {
            return View();
        }

        [HttpPost]
        public ActionResult Create([BindExcludeId] SexView sex)
        {
            if (!Validator.CanCreate(sex))
            {
                return View(sex);
            }

            Service.Create(sex);

            return RedirectToAction("Index");
        }

        [HttpGet]
        public ActionResult Details(int id)
        {
            return NotEmptyView(Service.Get<SexView>(id));
        }

        [HttpGet]
        public ActionResult Edit(int id)
        {
            return NotEmptyView(Service.Get<SexView>(id));
        }

        [HttpPost]
        public ActionResult Edit(SexView sex)
        {
            if (!Validator.CanEdit(sex))
            {
                return View(sex);
            }

            Service.Edit(sex);

            return RedirectToAction("Index");
        }

        [HttpGet]
        public ActionResult Delete(int id)
        {
            return NotEmptyView(Service.Get<SexView>(id));
        }

        [HttpPost]
        [ActionName("Delete")]
        public RedirectToActionResult DeleteConfirmed(int id)
        {
            if (!Validator.CanDelete(id))
            {
                return RedirectToAction("Delete", new { id });
            }

            Service.Delete(id);

            return RedirectToAction("Index");
        }
    }
}
